提取trackRoute的公用部分
