+++ src/components/TrackPageView.js
@@ -1,8 +1,7 @@
 import {Component} from 'react';
-import {Route, withRouter} from 'react-router-dom';
 import {omit} from 'lodash';
 
-class TrackPageView extends Component {
+export default class TrackPageView extends Component {
 
     componentDidMount() {
         this.trackPageView();
@@ -21,10 +20,9 @@ class TrackPageView extends Component {
     }
 
     render() {
-        const childProps = omit(this.props, ['location', 'match', 'history', 'tracker']);
+        const childProps = omit(this.props, ['location', 'match', 'history']);
+        const {ComponentIn} = this.props;
 
-        return <Route {...childProps} />;
+        return <ComponentIn {...childProps} />;
     }
 }
-
-export default withRouter(TrackPageView);
