+++ src/hocs/trackPageView.js
@@ -1,0 +1,8 @@
+import {Consumer} from '../components/TrackerContext';
+import TrackPageView from '../components/TrackPageView';
+
+export default ComponentIn => props => (
+    <Consumer>
+        {tracker => <TrackPageView {...props} ComponentIn={ComponentIn} tracker={tracker} />}
+    </Consumer>
+);
