+++ src/index.js
@@ -1,4 +1,4 @@
 export * from './components';
 export * from './providers';
 export * from './collect';
-export * from './hoc';
+export * from './hocs';
